+++ ignite/services/scaffolder/message.go
@@ -158,7 +158,7 @@ func (s Scaffolder) AddMessage(
 	if err != nil {
 		return sm, err
 	}
-	return sm, finish(ctx, cacheStorage, opts.AppPath, s.modpath.RawPath)
+	return sm, finish(ctx, cacheStorage, opts.AppPath, s.modpath.RawPath, false)
 }
 
 // checkForbiddenMessageField returns true if the name is forbidden as a message name.